+++ lib/src/controls/form/combo_box.dart
@@ -4,7 +4,6 @@ import 'dart:ui' show window;
 import 'package:flutter/foundation.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter/material.dart' as m;
 
 import 'package:fluent_ui/fluent_ui.dart';