+++ lib/src/controls/inputs/checkbox.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/rendering.dart';
-import 'package:flutter/material.dart' as m;
 
 import 'package:fluent_ui/fluent_ui.dart';